Debugging scripts

+++ src/docker/docker-build.sh
@@ -1,13 +1,16 @@
 #!/bin/bash
 
-# Ensure we are running in the proper folder
-[[ -d "./src/docker" ]] && [[ -f "./src/main.js" ]] || \
-     echo "This script must be run from the repository root folder"; exit 1
+echo "Ensure we are running in the proper folder"
+if !([[ -d "./src/docker" ]] && [[ -f "./src/main.js" ]]); then 
+    echo "This script must be run from the repository root folder"
+    exit 1
+fi
 
 # CI Will not have access to .gitignore'd files, so we remove node_modules
+echo "Removing modules"
 rm -r ./node_modules
 
-# Build Docker Image
+echo "Building Docker Image"
 docker build --file src/docker/Dockerfile --tag ghcr.io/agile-learning-institute/mentorhub-person-ui:latest .
 if [ $? -ne 0 ]; then
     echo "Docker build failed"
@@ -17,5 +20,5 @@ fi
 # Start the containers
 mh up person
 
-# Reinstall dependencies removed in the beginning of the script
-npm install
+echo "Restoring modules"
+npm install > /dev/null 2> /dev/null
